chore: replace Playwright with Puppeteer
Playwright was being a pain: running the test server was flakey and
switching to Puppeteer and Mocha gives us much more fine-grained control
over the test server and database setup steps. It's feeling a lot more
stable and I don't get intermittent failures where the test server fails
to be shut down properly.

This also allows us to assign a random available port rather than having
to specify one in the end-to-end tests.